Javav2-checksum example: Change int to long to support very large files (#6363)
Change int to long to support very large files

+++ javav2/example_code/s3/src/main/java/com/example/s3/BasicOpsWithChecksums.java
@@ -165,10 +165,10 @@ public class BasicOpsWithChecksums {
 
         try (RandomAccessFile file = new RandomAccessFile(filePath, "r")) {
             long fileSize = file.length();
-            int position = 0;
+            long position = 0;
             while (position < fileSize) {
                 file.seek(position);
-                int read = file.getChannel().read(bb);
+                long read = file.getChannel().read(bb);
 
                 bb.flip(); // Swap position and limit before reading from the buffer.
                 UploadPartRequest uploadPartRequest = UploadPartRequest.builder()
